Version 6.00

+++ phpsmk/restoran/kategori/select.php
@@ -23,7 +23,7 @@ $no = 1 + $mulai;
 </div>
 
 <h3>Kategori</h3>
-<table class="table table-bordered w-50">
+<table class="table table-bordered w-50 mt-4">
     <thead>
         <tr>
             <th>No</th>
@@ -34,18 +34,17 @@ $no = 1 + $mulai;
     </thead>
 
     <tbody>
-        <?php foreach ($row as $r) : ?>
+        <?php if (!empty($row)) { ?>
+            <?php foreach ($row as $r) : ?>
 
-            <tr>
-                <td><?php echo $no++ ?></td>
-                <td><?php echo $r['kategori'] ?></td>
-                <td><a href="?f=kategori&m=delete&id=<?php echo $r['idkategori'] ?>">Delete</a></td>
-                <td><a href="?f=kategori&m=update&id=<?php echo $r['idkategori'] ?>">Update</a></td>
-
-            </tr>
-
-        <?php endforeach ?>
-
+                <tr>
+                    <td><?php echo $no++ ?></td>
+                    <td><?php echo $r['kategori'] ?></td>
+                    <td><a href="?f=kategori&m=delete&id=<?php echo $r['idkategori'] ?>">Delete</a></td>
+                    <td><a href="?f=kategori&m=update&id=<?php echo $r['idkategori'] ?>">Update</a></td>
+                </tr>
+            <?php endforeach ?>
+        <?php } ?>
 
     </tbody>
 </table>
